% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_s3.R
\name{copy_s3}
\alias{copy_s3}
\title{Copy files from/to Amazon s3 drive.}
\usage{
copy_s3(x, from.folder, destination.folder, include, folder)
}
\arguments{
\item{x}{Filename inside the \code{from.folder} to copy}

\item{from.folder}{The folder where the files are stored}

\item{destination.folder}{The destination folder}

\item{include}{To include certain type of files}

\item{folder}{To transfer entire folder uses aws --recursive option}
}
\description{
Copy files from/to Amazon s3 drive.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}

